Keep algebra dispatch checks in dispatch lane
